# specs/003-inbox-home/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: CDevi API — Inbox Home Page (specs/003)
  version: 0.3.0
  description: Generated from Zod schemas in packages/contracts (`pnpm -F @cdevi/contracts openapi`) and snapshot-tested; edit the schemas, not the yaml. Users authenticate with a session cookie, ingestion principals with a Bearer token. Every error is a Problem. Every list is bounded to 50 items with keyset cursors.
servers:
  - url: /api
...
paths:
...
  /inbox:
    get:
      tags:
        - inbox
      summary: Inbox snapshot — counts, Today summary and one page of one tab (FR-006…FR-018, FR-026…FR-028)
      security:
        - sessionCookie: []
      parameters:
        - name: tab
          in: query
          schema:
            type: string
            enum:
              - needsYou
              - running
              - done
            default: needsYou
        - name: project
          in: query
          description: Project id or `all` (default). Must be visible to the user, else 404.
          schema:
            type: string
            default: all
        - name: cursor
          in: query
          description: Opaque keyset cursor from a previous nextCursor.
          schema:
            type: string
            maxLength: 200
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxSnapshot"
          headers:
            Server-Timing:
              schema:
                type: string
              description: db;dur=…, total;dur=…
        "400":
          description: Invalid query or cursor
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Project not visible
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /inbox/stream:
    get:
      tags:
        - inbox
      summary: Server-Sent Events — inbox.changed notifications for visible projects (FR-019, SC-007)
      description: 'text/event-stream. Each event: id=<seq>, event=inbox.changed, data={"projectId","workflowId"}. Heartbeat comment every 25 s. Send Last-Event-ID on reconnect to replay missed events (≤ 24 h) once. Payloads carry no titles or asks — the client refetches GET /inbox.'
      security:
        - sessionCookie: []
      parameters:
        - name: Last-Event-ID
          in: header
          schema:
            type: string
      responses:
        "200":
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /inbox/records/{kind}/{id}:
    get:
      tags:
        - inbox
      summary: Read-only record behind the /approvals/{id} and /workflows/{id} stubs (research R9)
      security:
        - sessionCookie: []
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum:
              - approvals
              - workflows
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordView"
        "401":
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found or not visible (403 and 404 are indistinguishable)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /ingest/workflows/{externalId}:
    put:
      tags:
        - ingest
      summary: Create or update a workflow; optionally apply a state transition (FR-020, FR-021)
      security:
        - ingestionToken: []
      parameters:
        - &a1
          name: externalId
          in: path
          required: true
          schema:
            type: string
            pattern: ^[A-Za-z0-9._:-]{1,128}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpsert"
      responses:
        "200":
          description: accepted, or stale when observedAt is not newer than the current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400": &a2
          description: Invalid body
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401": &a3
          description: Unknown or disabled principal
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403": &a4
          description: Project outside the principal scope
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404": &a5
          description: Unknown project or workflow
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Illegal transition for the current state
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /ingest/workflows/{externalId}/transitions:
    post:
      tags:
        - ingest
      summary: Append a state transition (state machine in data-model.md §3)
      security:
        - ingestionToken: []
      parameters:
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transition"
      responses:
        "200":
          description: accepted, or stale when observedAt is not newer than the current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400": *a2
        "401": *a3
        "403": *a4
        "404": *a5
        "409":
          description: Illegal transition for the current state (logged as rejected)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /ingest/approvals/{externalId}:
    put:
      tags:
        - ingest
      summary: Raise, update or resolve an approval request (FR-020, FR-021)
      description: A workflow may have at most one pending approval or clarification; a second pending request → 409.
      security:
        - ingestionToken: []
      parameters:
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalUpsert"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400": *a2
        "401": *a3
        "403": *a4
        "404": *a5
        "409":
          description: Another request is pending for this workflow
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /ingest/clarifications/{externalId}:
    put:
      tags:
        - ingest
      summary: Raise, update or answer a clarification (FR-020, FR-021)
      security:
        - ingestionToken: []
      parameters:
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClarificationUpsert"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400": *a2
        "401": *a3
        "403": *a4
        "404": *a5
        "409":
          description: Another request is pending for this workflow
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cdevi_session
    ingestionToken:
      type: http
      scheme: bearer
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
      required:
        - type
        - title
        - status
...
    InboxSnapshot:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
        project:
          type: string
        tab:
          type: string
          enum:
            - needsYou
            - running
            - done
        counts:
          type: object
          properties:
            needsYou:
              type: integer
              minimum: 0
              maximum: 9007199254740991
            running:
              type: integer
              minimum: 0
              maximum: 9007199254740991
            done:
              type: integer
              minimum: 0
              maximum: 9007199254740991
          required:
            - needsYou
            - running
            - done
        items:
          maxItems: 50
          type: array
          items:
            type: object
            properties:
              workflowId:
                type: string
                format: uuid
                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              title:
                type: string
              project:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  key:
                    type: string
                required:
                  - id
                  - key
              agent:
                nullable: true
                type: string
              state:
                type: string
                enum:
                  - QUEUED
                  - RUNNING
                  - RETRYING
                  - WAITING
                  - WAITING_FOR_HUMAN
                  - BLOCKED
                  - FAILED
                  - COMPLETED
                  - CANCELLED
              tab:
                type: string
                enum:
                  - needsYou
                  - running
                  - done
              kind:
                type: string
                enum:
                  - approval
                  - clarification
                  - blocked
                  - failed
                  - running
                  - done
              ask:
                nullable: true
                type: string
              riskLevel:
                nullable: true
                type: string
                enum:
                  - LOW
                  - MEDIUM
                  - HIGH
                  - CRITICAL
              raisedAt:
                type: string
                format: date-time
                pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
              isStale:
                type: boolean
              expiry:
                nullable: true
                type: object
                properties:
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  isExpired:
                    type: boolean
                required:
                  - expiresAt
                  - isExpired
              hasRecommendedAnswer:
                type: boolean
              stage:
                nullable: true
                type: object
                properties:
                  index:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  count:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  name:
                    nullable: true
                    type: string
                required:
                  - index
                  - count
                  - name
              startedAt:
                nullable: true
                type: string
                format: date-time
                pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
              finishedAt:
                nullable: true
                type: string
                format: date-time
                pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
              pullRequestRef:
                nullable: true
                type: string
              requestId:
                nullable: true
                type: string
                format: uuid
                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              href:
                type: string
            required:
              - workflowId
              - title
              - project
              - agent
              - state
              - tab
              - kind
              - ask
              - riskLevel
              - raisedAt
              - isStale
              - expiry
              - hasRecommendedAnswer
              - stage
              - startedAt
              - finishedAt
              - pullRequestRef
              - requestId
              - href
        nextCursor:
          nullable: true
          type: string
        today:
          type: object
          properties:
            timezone:
              type: string
            windowStart:
              type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            workflowsStarted:
              type: object
              properties:
                value:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                href:
                  type: string
              required:
                - value
                - href
            workflowsCompleted:
              type: object
              properties:
                value:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                href:
                  type: string
              required:
                - value
                - href
            approvalsDecided:
              type: object
              properties:
                value:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                href:
                  type: string
              required:
                - value
                - href
            needsYou:
              type: object
              properties:
                value:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                href:
                  type: string
              required:
                - value
                - href
          required:
            - timezone
            - windowStart
            - workflowsStarted
            - workflowsCompleted
            - approvalsDecided
            - needsYou
        policySummary:
          type: object
          properties:
            text:
              type: string
            href:
              type: string
          required:
            - text
            - href
      required:
        - generatedAt
        - project
        - tab
        - counts
        - items
        - nextCursor
        - today
        - policySummary
    RecordView:
      type: object
      properties:
        item:
          type: object
          properties:
            workflowId:
              type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            title:
              type: string
            project:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                key:
                  type: string
              required:
                - id
                - key
            agent:
              nullable: true
              type: string
            state:
              type: string
              enum:
                - QUEUED
                - RUNNING
                - RETRYING
                - WAITING
                - WAITING_FOR_HUMAN
                - BLOCKED
                - FAILED
                - COMPLETED
                - CANCELLED
            tab:
              type: string
              enum:
                - needsYou
                - running
                - done
            kind:
              type: string
              enum:
                - approval
                - clarification
                - blocked
                - failed
                - running
                - done
            ask:
              nullable: true
              type: string
            riskLevel:
              nullable: true
              type: string
              enum:
                - LOW
                - MEDIUM
                - HIGH
                - CRITICAL
            raisedAt:
              type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            isStale:
              type: boolean
            expiry:
              nullable: true
              type: object
              properties:
                expiresAt:
                  type: string
                  format: date-time
                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                isExpired:
                  type: boolean
              required:
                - expiresAt
                - isExpired
            hasRecommendedAnswer:
              type: boolean
            stage:
              nullable: true
              type: object
              properties:
                index:
                  type: integer
                  minimum: -9007199254740991
                  maximum: 9007199254740991
                count:
                  type: integer
                  minimum: -9007199254740991
                  maximum: 9007199254740991
                name:
                  nullable: true
                  type: string
              required:
                - index
                - count
                - name
            startedAt:
              nullable: true
              type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            finishedAt:
              nullable: true
              type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            pullRequestRef:
              nullable: true
              type: string
            requestId:
              nullable: true
              type: string
              format: uuid
              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            href:
              type: string
          required:
            - workflowId
            - title
            - project
            - agent
            - state
            - tab
            - kind
            - ask
            - riskLevel
            - raisedAt
            - isStale
            - expiry
            - hasRecommendedAnswer
            - stage
            - startedAt
            - finishedAt
            - pullRequestRef
            - requestId
            - href
        resolution:
          nullable: true
          type: object
          properties:
            outcome:
              type: string
              enum:
                - approved
                - rejected
                - answered
            at:
              type: string
              format: date-time
              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
            by:
              nullable: true
              type: string
          required:
            - outcome
            - at
            - by
      required:
        - item
        - resolution
    WorkflowUpsert:
      type: object
      properties:
        projectKey:
          type: string
          minLength: 1
          maxLength: 64
        title:
          type: string
          minLength: 1
          maxLength: 200
        agent:
          nullable: true
          type: string
          minLength: 1
          maxLength: 80
        state:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - RETRYING
            - WAITING
            - WAITING_FOR_HUMAN
            - BLOCKED
            - FAILED
            - COMPLETED
            - CANCELLED
        reason:
          nullable: true
          type: string
          minLength: 1
          maxLength: 240
        stage:
          nullable: true
          type: object
          properties:
            index:
              type: integer
              minimum: 1
              maximum: 9007199254740991
            count:
              type: integer
              minimum: 1
              maximum: 20
            name:
              nullable: true
              type: string
              minLength: 1
              maxLength: 60
          required:
            - index
            - count
        pullRequestRef:
          nullable: true
          type: string
          minLength: 1
          maxLength: 200
        observedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
      required:
        - projectKey
        - title
        - observedAt
...